raspberry-pi/requirements.txt
sparkfun-qwiic-otos
pyserial
cobs